Split train_llm into per-loop modules under training/llm.
