+++ demos/llm/debugging/debugging_llm_stage_2.py
@@ -18,7 +18,7 @@ from llm_debug_utils import lora_config_from_dict
 from tiny_model import TinyDigitTokenizer, build_tiny_actor_network
 
 from agilerl.algorithms import GRPO, LLMPPO, LLMREINFORCE
-from agilerl.training import train_llm
+from agilerl.training.llm import multiturn as train_llm
 from agilerl.training.train_llm import finetune_llm_multiturn
 from agilerl.utils.llm_utils import create_llm_accelerator, masked_whiten
 from agilerl.utils.probe_envs_llm import MultiInputConditionalEnv
